+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- A static IsNullOrEmpty method, to check if a Range<T> is null or Range<T>.Empty
+- The Merge method, to create a single Range<T> from two overlapping or touching Ranges
+- The Split method, to split the Range<T> at a given value
+- The Intersection method, that returns the Range<T> that is part of both Ranges
+
+### Changed
+- The implementation of Equals and CompareTo so that every Empty Range, ranges with the same start and end, are equal regarding of the value of start and end
+- The implementation of Includes, Overlaps, Touches and Gap in combination with the current Range<T> bieng Empty and/or the specified Range<T> bieng Empty
+
 ## [0.1.4] - 2018-12-23
 ### Added
 - The Includes(T value) method was added to Range<T>, it returns true if the given value is a part of the Range
@@ -34,5 +45,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   standardized open source project CHANGELOG.
 
 [Unreleased]: https://github.com/reynj/reynj/compare/v0.1.4...HEAD
-[0.1.4]: https://github.com/reynj/reynj/compare/0.1.2...v0.1.4
-[0.1.3]: https://github.com/reynj/reynj/compare/0.1.2...0.1.3
+[0.1.4]: https://github.com/reynj/reynj/compare/v0.1.2...v0.1.4
+[0.1.3]: https://github.com/reynj/reynj/compare/v0.1.2...v0.1.3